backend_currency_converter/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.0
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==3.13.0
flake8==7.1.1
gunicorn==20.1.0
idna==3.10
mccabe==0.7.0
pillow==11.0.0
psycopg2-binary==2.9.3
pycodestyle==2.12.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.2.3
